Fix: Please use the already existing WidgetController.php
