+++ templates/index.html.twig
@@ -96,7 +96,7 @@
 
                 <li class="nav-item">
                     <a class="nav-link ajax-link" href="{{ path('widgets_index') }}">
-                        <i class="fa fa-plus"></i>
+                        <i class="fa fa-cogs"></i>
                         Manage Widgets
                     </a>
                 </li>
